add sample code for code analysis [#167956173]
