+++ aspnet/HousingXyz.Provider.Api/Controllers/RoomController.cs
@@ -1,0 +1,45 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+
+namespace HousingXyz.Provider.Api.Controllers
+{
+  [Route("api/[controller]")]
+  [ApiController]
+  public class RoomController : ControllerBase
+  {
+    // GET api/values
+    [HttpGet]
+    public ActionResult<IEnumerable<string>> Get()
+    {
+      return new string[] { "room1", "room2" };
+    }
+
+    // GET api/values/5
+    [HttpGet("{id}")]
+    public ActionResult<string> Get(int id)
+    {
+      return "room";
+    }
+
+    // POST api/values
+    [HttpPost]
+    public void Post([FromBody] string value)
+    {
+    }
+
+    // PUT api/values/5
+    [HttpPut("{id}")]
+    public void Put(int id, [FromBody] string value)
+    {
+    }
+
+    // DELETE api/values/5
+    [HttpDelete("{id}")]
+    public void Delete(int id)
+    {
+    }
+  }
+}
